+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.0.28] - 2021-06-XX
 
+### Added
+
+* [Add Jenkins CI integration](https://docs.restqa.io/ci-cd/jenkins) | [#105](https://github.com/restqa/restqa/pull/105)
+
 ## [0.0.27] - 2021-05-29
 
 ### Added